# specs/1-dapr-integration/contracts/todo-api-openapi.yaml
openapi: 3.0.0
info:
  title: Todo Application Backend API with Dapr Integration
  version: 1.0.0
  description: API for the Todo Application with Dapr integration for pub/sub, state management, and service invocation

servers:
  - url: http://localhost:8000
    description: Local development server with Dapr sidecar

paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks with optional filtering
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      description: Create a new task and publish a creation event via Dapr pub/sub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'

  /tasks/{taskId}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID from state store via Dapr
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a task
      description: Update an existing task and publish an update event via Dapr pub/sub
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a task
      description: Delete a task and publish a deletion event via Dapr pub/sub
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /dapr-subscribe:
    get:
      summary: Dapr subscription endpoint
      description: Endpoint for Dapr to discover which topics this service wants to subscribe to
      responses:
        '200':
          description: Subscription configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    pubsubname:
                      type: string
                      example: "kafka-pubsub"
                    topic:
                      type: string
                      example: "task-updates"
                    route:
                      type: string
                      example: "/tasks/update"

  /reminders:
    post:
      summary: Schedule a reminder
      description: Schedule a reminder and store it in state via Dapr
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reminder'
      responses:
        '201':
          description: Reminder scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'

  /process-reminders:
    post:
      summary: Process scheduled reminders (called by cron binding)
      description: Endpoint called by Dapr cron binding to process due reminders
      responses:
        '200':
          description: Reminders processed successfully
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    Task:
      type: object
      required:
        - title
      properties:
        taskId:
          type: string
          description: Unique identifier for the task
          example: "task-123"
        title:
          type: string
          description: Title of the task
          example: "Complete project proposal"
        description:
          type: string
          description: Detailed description of the task
          example: "Finish the project proposal document and send for review"
        status:
          type: string
          enum: [pending, in-progress, completed]
          description: Current status of the task
          example: "pending"
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level of the task
          example: "high"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when task was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when task was last updated
        dueDate:
          type: string
          format: date-time
          description: Due date for the task
        recurrenceRule:
          type: string
          description: Rule for recurring tasks
          example: "FREQ=DAILY;INTERVAL=1"
    
    Reminder:
      type: object
      required:
        - taskId
        - reminderTime
      properties:
        reminderId:
          type: string
          description: Unique identifier for the reminder
          example: "reminder-456"
        taskId:
          type: string
          description: Reference to the associated task
          example: "task-123"
        reminderTime:
          type: string
          format: date-time
          description: Time when reminder should be triggered
        notified:
          type: boolean
          description: Flag indicating if reminder has been sent
          default: false
        createdAt:
          type: string
          format: date-time
          description: Timestamp when reminder was created

  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid input data"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "An unexpected error occurred"